pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aiforge-engine"
version = "0.0.11"
description = "Intelligent Intent Adaptive Execution Engine - 智能意图自适应执行引擎"
readme = "README_EN.md"
license = { text = "Apache 2.0" }
requires-python = ">=3.10"

authors = [{ name = "iniwap", email = "iniwaper@gmail.com" }]
maintainers = [{ name = "iniwap", email = "iniwaper@gmail.com" }]

keywords = [
    "ai",
    "automation",
    "intent",
    "execution",
    "engine",
    "llm",
    "natural-language",
    "code-generation",
    "search",
    "aipyapp",
    "crewai",
    "agent",
    "ai-agents",
    "manus",
    "ai-tools",
]

classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Code Generators",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
    "Operating System :: OS Independent",
    "Environment :: Console",
    "Environment :: Web Environment",
]

dependencies = [
    "requests>=2.25.0",
    "beautifulsoup4>=4.9.0",
    "click>=8.0.0",
    "tomlkit>=0.10.0",
    "peewee>=3.14.0",
    "pydantic>=2.0.0",
    "typing-extensions>=4.0.0",
    "rich>=12.0.0",
    "setuptools",
    "python-dateutil>=2.8.0",
]

[project.optional-dependencies]
dev = ["python-dotenv>=1.0.0", "pytest>=7.0.0"]
transformers = ["sentence-transformers>=2.0.0"]
web = ["fastapi>=0.100.0", "uvicorn>=0.20.0", "jinja2>=3.1.0"]
gui = [
    "pywebview>=4.0.0",
    "pystray>=0.19.0",
    "pillow>=8.0.0",
    "pywin32>=306;platform_system=='Windows'",
]
deploy = ["docker>=6.0.0", "pyyaml>=6.0.0"]
all = [
    "sentence-transformers>=2.0.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
    "jinja2>=3.1.0",
    "pywebview>=4.0.0",
    "docker>=6.0.0",
    "pyyaml>=6.0.0",
]

[project.urls]
Homepage = "https://github.com/iniwap/AIForge"
Documentation = "https://iniwap.github.io/AIForge"
Repository = "https://github.com/iniwap/AIForge"
"Bug Tracker" = "https://github.com/iniwap/AIForge/issues"
Changelog = "https://github.com/iniwap/AIForge/releases"
"Source Code" = "https://github.com/iniwap/AIForge"
Discord = "https://discord.gg/Vp35uSBsrw"
Twitter = "https://twitter.com/iafun_tipixel"

[project.scripts]
aiforge = "aiforge.cli:main"

[tool.hatch.build.targets.wheel]
packages = [
    "src/aiforge",
    "src/aiforge_web",
    "src/aiforge_gui",
    "src/aiforge_deploy",
]

[tool.hatch.build.targets.sdist]
include = [
    "/src/aiforge",
    "/src/aiforge_web",
    "/src/aiforge_gui",
    "/src/aiforge_deploy",
    "/README.md",
    "/LICENSE",
    "/pyproject.toml",
]

[dependency-groups]
dev = ["python-dotenv>=1.1.1"]

[tool.pytest.ini_options]
pythonpath = ["src"]
